Split native-structure tests into 3 test files
3 configurations:
* (always)
* codegen-full
* codegen-full && experimental-threads
